Avoid double launch of setupLists () The old grammar "Consumer" is excluded from the lambda expression

+++ app/src/main/java/com/example/camera2probeK/CameraSpecResult.kt
@@ -1,5 +1,5 @@
 // Toyoaki, OHGOCHI  https://twitter.com/Ohgochi/
 package com.example.camera2probeK
 
-class CameraSpecResult internal constructor(t1: String, t2: String, t3: Int) :
+class CameraSpecResult internal constructor(t1: String, t2: String, t3: Float) :
     Triple<Any, Any, Any?>(t1, t2, t3)
